Agregada validacion para no eliminar cursos con alumnos

+++ src/components/Course.js
@@ -64,9 +64,6 @@ const styles = {
 const Course = ({ courseData, removeCourse }) => {
   const localUser = JSON.parse(localStorage.getItem('current-user'));
   const currentUser = useContext(UserContext);
-  console.log('courseData', courseData);
-  console.log('currentUser', currentUser);
-  console.log('localUser', localUser);
   const {
     id,
     name,
@@ -123,18 +120,22 @@ const Course = ({ courseData, removeCourse }) => {
   };
 
   const handleRemoveCourse = () => {
-    const answer = window.confirm(
-      '¿Estás seguro/a de que quieres realizar esta acción?'
-    );
-    if (answer) {
-      removeCourse();
-      fetch(`${endpoint}/courses/${id}`, {
-        method: 'DELETE',
-        headers: {
-          'Content-type': 'application/json',
-          Authorization: `Bearer ${localUser?.token}`,
-        },
-      });
+    if (inscriptions.length) {
+      alert('No puedes eliminar un curso con alumnos inscriptos.');
+    } else {
+      const answer = window.confirm(
+        '¿Estás seguro/a de que quieres realizar esta acción?'
+      );
+      if (answer) {
+        removeCourse();
+        fetch(`${endpoint}/courses/${id}`, {
+          method: 'DELETE',
+          headers: {
+            'Content-type': 'application/json',
+            Authorization: `Bearer ${localUser?.token}`,
+          },
+        });
+      }
     }
   };
